Protein Function Prediction Tool
Proteins are annotated with functionalities based on different annotations techniques (sequence similarity to another protein, experimental validation, etc.). These functionalities are represented as terms (e.g., GO:00000219) from Gene Ontology. In this project, the aim is  predicting new functionalities for proteins based on the currently assigned functionalities for that protein.